type clean up
